Separate GEM contraction planning
